requirements.txt
scholarly>=1.7.0
pyyaml>=6.0

